Refactor relationship filters (#10)
Refactor relationship filters to be interms of entities instead of attributes. Rely on underlay resolution to find the entities to use for creating queries. This better reflects the logical model and will let the API operate in terms of entities without knowing what attributes are used to relate entities.
Refactor the UnderlaySqlResolver to make it a direct part of SqlVisitor with a dependency on the Underlay. This keeps the particulars of SQL generation to the appropriate search entity.