Add initial index job for flattening hierachical relationships (#36)
Adds an Apache Beam job for creating a BQ table as an index for a flattened hierachical relationship of ancestor -> all descendants.
Provides some notes on an end-end working example in the broad-tanagra-dev project.
This does not begin to address making this easy to make further deployments of. I'd like to have a few more examples before creating some common ways to make these easeir to re-use.