Add a SearchTermFilter to the API for text search on entities. (#40)
This is meant to allow the text search for things like concepts.
This is a somewhat unusual filter in that we maybe don't want to encourage users to include it in their cohort building, but it's mostly used to select entities during cohort construction.
Also note that this will probably only be allowed to use in retrieving entities, but not in BQ index generation as we may not have an equivalent BQ text search operation.

We could do another API approach, like having a dedicated instances:searchTerm API and not include SearchTermFilter as a part of a general Filter. That new API may look a lot like instances:search as we want to return values for entity instances. that would make it clearer that this a special use case, as opposed to generic entity filtering. I'm ambivalent.
LMK if a 1 pager discussing this would be helpful.

TODOs for support for this:
Underlay configuration support
Text search indexes for entities
Reconsider how EntitySearchHint should relate to this.