Generalize criteria data updating (#123)
This moves the actual data updating out of the criteria which will allow
criteria to be used for bor cohorts and concept sets.

concept.test.tsx is a bit of a mess. It was creating cohorts for each
test case but only using the first one. I think the solution will be to
split out some of the logic and test that with unit tests, while leaving
most of the UI testing to a broader level. That aligns with Redux's
testing suggestions but I need to think about it some more.