Add basic local storage support
Storage is implemented by storing whenever the Redux store changes and
loading on app startup using a plugin to allow the implementation to be
swapped out. Depending on requirements goingn forwards, this may need to
be expanded to support editing individual objects, watching for updates
and conflict resolution, offline editing, etc.