Avoid refetching counts for unrelated changes (#196)
* Separate criteria plugin config and data

Since the config was stored with the data, loading saved criteria would
continue to used outdated config. This caused issues like not displaying
new columns added to the config. Instead, only the id of the config is
stored in the criteria and the current version is passed into the
plugins.

This also merges the plugin specific config into the same structure as
the rest of the config. The separation meant plugins that wanted base
config and custom config either had to store them separately or cast the
custom data at every use.

* Avoid refetching counts for unrelated changes