Refactor ActionBar to be more customizable
I tried various more centralized approaches but ran into state or
routing issues with each one. Instead, this applies an overall layout
and each "page" is responsible for show it's own ActionBar. There are a
few slightly tricky parts and a bit of duplicated code but both are
minor and it's straightforward overall.

* Top level layout allocates space for the ActionBar on all "pages".
* No more duplicate ActionBars stacked on top of each other.
* ActionBar can be customized for different views, including upcoming
  cohort review.
* Refactor criteria based views as the amount of duplicated code
  increases.