Migrate from useAsync to SWR
SWR is a more robust library for handling web requests and will offer
support for automatically revalidating data (e.g. when adding a cohort
and refetching the list). and caching. Migrating the existing code first
so that there won't be two libraries once artifact storage support is
added.

Each invocation of useSWR requires a "key" object for cache management
purposes. A future PR could consider moving some of the parameters out
of the useCallback list since the key is passed to the fetch function
but in many cases that doesn't make sense because the object (e.g. an
underlay) shouldn't be included in the key in its entirety so some
further thought is required.

Some places in the code that were managing fetched data in a separate
state variable have been converted to use the data out of the fetch
state instead. This enables caching to work correctly, eliminates some
weird corner cases, and is more straightforward.

Response error parsing has been moved into the sourse since the existing
method is incompatible with SWR. It makes more sense for it to be there
anyways.