Show top results for empty queries instead of root items
Unlike AoU, this currently leaves the root nodes visible in the list
view but the UI can't globablly exclude them because they should be
visible in some places (e.g. genotyping platform). Configuration for
that behavior can be added in a future PR but given the varying behavior
of top level nodes in AoU hierarchies it may go beyond a UI config flag.