Add new visualization support using Observable
This is pluggable framework for visualizations that currently just
replaces the demographics charts. This gives us something to hook up the
new queries and configuration to as we build it out.

Also fixes an issue that was causing empty cohorts to unnecessarily
refresh the demographic charts. A new random id was being generated
every time an empty cohort was fetched causing the UI to constantly
think it had changed.

After experiments with other charting libraries didn't pan out, this
largely leaves the existing demographic chart code intact but moved into
a plugin that's instantiated for each chart.