Add support for non-person visualizations
Update sorting viz options since the attribute is already present but
the value needs to be an option.

This support is not fully generic since the UI no longer know which
occurrences to query for a given selector. For now the references will
be hard coded until M2 when the backend takes over viz query generation.