Fix selecting inside inline criteria controls deselecting the criteria
Dragging to select the value in inline criteria controls (e.g. age)
would cause the criteria to be selected on mouse button up. This
prevents that specific case from from happening but there's an
unavoidable case where the button down is in the text box and the button
up is elsewhere in the criteria. This is part of the way browsers
generate click events so we can't really do anything about it. We'll
have to consider alternative UXs (e.g. a more standard accordion with an
arrow).